relocates quill utils
