+++ packages/commonwealth/server/scripts/emails.ts
@@ -1,8 +1,8 @@
 import { logger } from '@hicommonwealth/core';
-import type {
-  AddressAttributes,
-  DB,
-  UserAttributes,
+import {
+  type AddressAttributes,
+  type DB,
+  type UserAttributes,
 } from '@hicommonwealth/model';
 import {
   DynamicTemplate,
@@ -11,17 +11,14 @@ import {
   ISnapshotNotificationData,
   NotificationCategories,
   getThreadUrl,
+  renderQuillDeltaToText,
 } from '@hicommonwealth/shared';
 import sgMail from '@sendgrid/mail';
 import _ from 'lodash';
 import { Op, WhereOptions } from 'sequelize';
 import { Label as ChainEventLabel } from '../../shared/chain/labelers/util';
 import type { CWEvent } from '../../shared/chain/types/types';
-import {
-  formatAddressShort,
-  renderQuillDeltaToText,
-  smartTrim,
-} from '../../shared/utils';
+import { formatAddressShort, smartTrim } from '../../shared/utils';
 import { config } from '../config';
 
 const log = logger(import.meta);
@@ -53,12 +50,12 @@ const getForumNotificationCopy = async (
     category_id === NotificationCategories.NewComment
       ? `Comment on: ${decodedTitle}`
       : category_id === NotificationCategories.NewMention
-      ? `You were mentioned in: ${decodedTitle}`
-      : category_id === NotificationCategories.NewCollaboration
-      ? `You were added as a collaborator on: ${decodedTitle}`
-      : category_id === NotificationCategories.NewThread
-      ? `New thread: ${decodedTitle}`
-      : 'New activity on Commonwealth';
+        ? `You were mentioned in: ${decodedTitle}`
+        : category_id === NotificationCategories.NewCollaboration
+          ? `You were added as a collaborator on: ${decodedTitle}`
+          : category_id === NotificationCategories.NewThread
+            ? `New thread: ${decodedTitle}`
+            : 'New activity on Commonwealth';
 
   // author
   // @ts-expect-error StrictNullChecks
@@ -99,15 +96,15 @@ const getForumNotificationCopy = async (
   ].includes(category_id)
     ? 'commented on'
     : category_id === NotificationCategories.NewMention
-    ? 'mentioned you in the thread'
-    : category_id === NotificationCategories.NewCollaboration
-    ? 'invited you to collaborate on'
-    : [
-        NotificationCategories.ThreadEdit,
-        NotificationCategories.NewThread,
-      ].includes(category_id)
-    ? 'created a new thread'
-    : null;
+      ? 'mentioned you in the thread'
+      : category_id === NotificationCategories.NewCollaboration
+        ? 'invited you to collaborate on'
+        : [
+              NotificationCategories.ThreadEdit,
+              NotificationCategories.NewThread,
+            ].includes(category_id)
+          ? 'created a new thread'
+          : null;
   const objectCopy = decodeURIComponent(root_title).trim();
   const communityObject = await models.Community.findOne({
     where: { id: community_id },
