+++ packages/commonwealth/shared/utils.ts
@@ -56,103 +56,6 @@ export const urlHasValidHTTPPrefix = (url: string) => {
   return url.indexOf('http://') === 0 || url.indexOf('https://') === 0;
 };
 
-export const preprocessQuillDeltaForRendering = (nodes) => {
-  // split up nodes at line boundaries
-  const lines = [];
-  for (const node of nodes) {
-    if (typeof node.insert === 'string') {
-      const matches = node.insert.match(/[^\n]+\n?|\n/g);
-      (matches || []).forEach((line) => {
-        // @ts-expect-error StrictNullChecks
-        lines.push({ attributes: node.attributes, insert: line });
-      });
-    } else {
-      // @ts-expect-error StrictNullChecks
-      lines.push(node);
-    }
-  }
-  // group nodes under parents
-  const result = [];
-  let parent = { children: [], attributes: undefined };
-  for (const node of lines) {
-    // @ts-expect-error StrictNullChecks
-    if (typeof node.insert === 'string' && node.insert.endsWith('\n')) {
-      // @ts-expect-error StrictNullChecks
-      parent.attributes = node.attributes;
-      // concatenate code-block node parents together, keeping newlines
-      if (
-        result.length > 0 &&
-        // @ts-expect-error StrictNullChecks
-        result[result.length - 1].attributes &&
-        parent.attributes &&
-        parent.attributes['code-block'] &&
-        // @ts-expect-error StrictNullChecks
-        result[result.length - 1].attributes['code-block']
-      ) {
-        // @ts-expect-error StrictNullChecks
-        parent.children.push({ insert: node.insert });
-        // @ts-expect-error StrictNullChecks
-        result[result.length - 1].children = result[
-          result.length - 1
-          // @ts-expect-error StrictNullChecks
-        ].children.concat(parent.children);
-      } else {
-        // @ts-expect-error StrictNullChecks
-        parent.children.push({ insert: node.insert });
-        // @ts-expect-error StrictNullChecks
-        result.push(parent);
-      }
-      parent = { children: [], attributes: undefined };
-    } else {
-      parent.children.push(node);
-    }
-  }
-  // If there was no \n at the end of the document, we need to push whatever remains in `parent`
-  // onto the result. This may happen if we are rendering a truncated Quill document
-  if (parent.children.length > 0) {
-    // @ts-expect-error StrictNullChecks
-    result.push(parent);
-  }
-
-  // trim empty newlines at end of document
-  while (
-    result.length &&
-    // @ts-expect-error StrictNullChecks
-    result[result.length - 1].children.length === 1 &&
-    // @ts-expect-error StrictNullChecks
-    typeof result[result.length - 1].children[0].insert === 'string' &&
-    // @ts-expect-error StrictNullChecks
-    result[result.length - 1].children[0].insert === '\n' &&
-    // @ts-expect-error StrictNullChecks
-    result[result.length - 1].children[0].attributes === undefined
-  ) {
-    result.pop();
-  }
-
-  return result;
-};
-
-export const renderQuillDeltaToText = (delta, paragraphSeparator = '\n\n') => {
-  return preprocessQuillDeltaForRendering(delta.ops)
-    .map((parent) => {
-      // @ts-expect-error StrictNullChecks
-      return parent.children
-        .map((child) => {
-          if (typeof child.insert === 'string')
-            return child.insert.trimRight('\n');
-          if (child.insert?.image) return '(image)';
-          if (child.insert?.twitter) return '(tweet)';
-          if (child.insert?.video) return '(video)';
-          return '';
-        })
-        .filter((child) => !!child)
-        .join(' ')
-        .replace(/  +/g, ' '); // remove multiple spaces
-    })
-    .filter((parent) => !!parent)
-    .join(paragraphSeparator);
-};
-
 export function formatAddressShort(
   address: string,
   chain?: string,
